see the comment on the line above what i changed :)
There was MASSIVE lagging with your remove function for the bloods
collection. like it just about broke the game when i tried to remove
lots at once, so i changed it to splice and it works fine now. I left
the remove function commented out just in case you wanted it still, but
splice is far more efficient.